Move RequestExceptions to exceptions.py
